+++ server/src/main/java/org/opensearch/index/IndexService.java
@@ -379,6 +379,7 @@ public class IndexService extends AbstractIndexComponent implements IndicesClust
             client,
             queryCache,
             directoryFactory,
+            null,
             remoteDirectoryFactory,
             eventListener,
             wrapperFactory,